muddy-test
Enables FEATURES="test" on all packages in the worldfile,
then enableson @system.
Done as a two pass when MP gets a feature to add packages via a script.

The script will add the common USE flags needed to get testing working
on stage3 packages, then adds a check to warn the user if the worldfile
will fail. The logic is basic for noow but works.

Usage: muddy-test /path/to/chroot

Signed-off-by: immoloism@gmail.com <immoloism@gmail.com>